requirements.txt
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
direnv==2020.12.3
dnspython==2.3.0
email-validator==1.3.1
env-file==2020.12.3
flake8==6.0.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.2.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mccabe==0.7.0
mysql==0.0.3
mysqlclient==2.2.4
optional-django==0.1.0
packaging==23.2
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.7.0
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
values==2020.12.3
Werkzeug==3.0.3
WTForms==3.0.1
zipp==3.19.2
setuptools==71.1.0
